refactor: Extract calculateACB into src/acb.ts
Moves the ACB calculation (including isSuperficialLoss, getPrice, and
the local getAsset helper) out of server.ts into a dedicated module.
server.ts now imports calculateACB and calls it from the /api/acb
handler. No behavioral changes.

Also removes a stray console.log(items) in getAsset that was shipping
to production.
